Bloody dates...

+++ AppointmentManager/Models/WorkingDayCoreDataManager.swift
@@ -13,7 +13,7 @@ class WorkingDayCoreDataManager: CoreDataManager {
     
     func createWorkingDay(from dayScheduleView: DayScheduleStackView) -> WorkingDay {
         let workingDay = WorkingDay(context: managedObjectContext)
-        workingDay.day = dayScheduleView.day
+        workingDay.day = Int16(dayScheduleView.day)
         workingDay.startingHour = dayScheduleView.startingHourStackView?.selectedDate
         workingDay.endingHour = dayScheduleView.endingHourStackView?.selectedDate
         return workingDay
